Move operataor options out to a sepearte public file
