+++ python/ray/data/__init__.py
@@ -71,6 +71,7 @@ from ray.data.read_api import (  # noqa: F401
     read_videos,
     read_webdataset,
 )
+from operation_options import OperatorOptions
 
 # Module-level cached global functions for callable classes. It needs to be defined here
 # since it has to be process-global across cloudpickled funcs.
@@ -175,4 +176,5 @@ __all__ = [
     "read_webdataset",
     "Preprocessor",
     "TFXReadOptions",
+    "OperatorOptions",
 ]
