+++ python/ray/data/_internal/execution/operators/actor_pool_map_operator.py
@@ -26,7 +26,6 @@ from ray.data._internal.execution.interfaces import (
     PhysicalOperator,
     RefBundle,
     TaskContext,
-    OperatorOptions,
 )
 from ray.data._internal.execution.interfaces.physical_operator import _ActorPoolInfo
 from ray.data._internal.execution.node_trackers.actor_location import (
@@ -41,6 +40,7 @@ from ray.data.block import Block, BlockMetadata
 from ray.data.context import DataContext
 from ray.types import ObjectRef
 from ray.util.common import INT32_MAX
+from ray.data.operation_options import OperatorOptions
 
 logger = logging.getLogger(__name__)
