+++ python/ray/data/_internal/logical/operators/map_operator.py
@@ -9,7 +9,7 @@ from ray.data._internal.logical.operators.one_to_one_operator import AbstractOne
 from ray.data.block import UserDefinedFunction
 from ray.data.expressions import Expr
 from ray.data.preprocessor import Preprocessor
-from ray.data._internal.execution.interfaces import OperatorOptions
+from ray.data.operation_options import OperatorOptions
 
 if TYPE_CHECKING:
     import pyarrow as pa
