+++ python/ray/data/_internal/planner/plan_udf_map_op.py
@@ -26,7 +26,7 @@ from ray._common.utils import get_or_create_event_loop
 from ray._private.ray_constants import env_integer
 from ray.data._expression_evaluator import eval_expr
 from ray.data._internal.compute import get_compute
-from ray.data._internal.execution.interfaces import PhysicalOperator, OperatorOptions
+from ray.data._internal.execution.interfaces import PhysicalOperator
 from ray.data._internal.execution.interfaces.task_context import TaskContext
 from ray.data._internal.execution.operators.map_operator import MapOperator
 from ray.data._internal.execution.operators.map_transformer import (
@@ -61,6 +61,7 @@ from ray.data.block import (
 from ray.data.context import DataContext
 from ray.data.exceptions import UserCodeException
 from ray.util.rpdb import _is_ray_debugger_post_mortem_enabled
+from ray.data.operation_options import OperatorOptions
 
 logger = logging.getLogger(__name__)
